refactor: プロフィール画面の共通化
